Revert "Use origami navigation service (behind a flag)"
