--- header/partials/drawer/template_v1.html
@@ -1,122 +1,0 @@
-<div class="o-header__drawer" id="o-header-drawer" role="navigation" aria-label="Drawer menu" data-o-header-drawer data-o-header-drawer--no-js data-trackable="drawer" data-trackable-terminate>
-	<div class="o-header__drawer-inner">
-
-		<div class="o-header__drawer-tools">
-			<a class="o-header__drawer-tools-logo" href="/" data-trackable="logo">
-				<span class="o-header__visually-hidden">Financial Times</span>
-			</a>
-
-			<button type="button" class="o-header__drawer-tools-close" aria-controls="o-header-drawer" title="Close drawer menu" data-trackable="close">
-				<span class="o-header__visually-hidden">Close</span>
-			</button>
-			{{#if @root.editions}}
-			<p class="o-header__drawer-current-edition">{{@root.editions.current.name}} Edition</p>
-			{{/if}}
-		</div>
-
-		<div class="o-header__drawer-search">
-			<form class="o-header__drawer-search-form" action="/search" role="search" aria-label="Site search" data-typeahead data-typeahead-categories="tags,equities" data-typeahead-view-all>
-				<label class="o-header__visually-hidden" for="o-header-drawer-search-term">Search the <abbr title="Financial Times">FT</abbr></label>
-				<input class="o-header__drawer-search-term" id="o-header-drawer-search-term" name="q" type="text" autocomplete="off" autocorrect="off" autocapitalize="off" spellcheck="false" placeholder="Search the FT" data-trackable="search-term" data-typeahead-input>
-				<button class="o-header__drawer-search-submit" type="submit" data-trackable="search-submit">
-					<span class="o-header__visually-hidden">Search</span>
-				</button>
-			</form>
-		</div>
-
-		<nav class="o-header__drawer-menu o-header__drawer-menu--primary o-header__drawer-homepage">
-
-
-			<ul class="o-header__drawer-menu-list">
-
-				{{#if @root.editions}}
-				{{#each @root.editions.others}}
-				<li class="o-header__drawer-menu-item" data-trackable="edition-switcher">
-					<a class="o-header__drawer-menu-link" href="/?edition={{id}}" data-trackable="{{id}}">Switch to {{name}} Edition</a>
-				</li>
-				{{/each}}
-				{{/if}}
-
-				{{#if @root.flags.compactView}}
-				{{#ifEquals @root.__name "front-page"}}
-				{{#ifEquals @root.viewStyle "compact"}}
-				<li class="o-header__drawer-menu-item" data-trackable="viewtoggle">
-					<a class="o-header__drawer-menu-link" href="viewtoggle/standard" data-trackable="standard">Turn off compact homepage</a>
-				</li>
-				{{else}}
-				<li class="o-header__drawer-menu-item">
-					<a class="o-header__drawer-menu-link" href="/viewtoggle/compact" data-trackable="compact">Turn on compact homepage</a>
-				</li>
-				{{/ifEquals}}
-				{{/ifEquals}}
-				{{/if}}
-			</ul>
-
-			<ul class="o-header__drawer-menu-list">
-				{{#each @root.navigationLists.drawer}}
-				{{#each this}}
-				{{#if heading}}
-				<li class="o-header__drawer-menu-item o-header__drawer-menu-item--heading">
-					{{heading}}
-				</li>
-				{{else}}
-				<li class="o-header__drawer-menu-item {{#unless @index}}o-header__drawer-menu-item--divide{{/unless}}">
-					{{#if children}}
-					<div class="o-header__drawer-menu-toggle-wrapper">
-						{{#with item}}
-						<a class="o-header__drawer-menu-link o-header__drawer-menu-link--parent o-header__drawer-menu-link--{{#if selected}}selected{{else}}unselected{{/if}}" href="{{href}}" {{#selected}}aria-label="Current page" aria-selected="true"{{/selected}} data-trackable="{{name}}">{{name}}</a>
-						<button class="o-header__drawer-menu-toggle o-header__drawer-menu-toggle--{{#if selected}}selected{{else}}unselected{{/if}}" aria-controls="o-header-drawer-child-{{@index}}" data-trackable="sub-level-toggle | {{name}}">
-							Show more {{name}} links
-						</button>
-						{{/with}}
-					</div>
-					<ul class="o-header__drawer-menu-list o-header__drawer-menu-list--child" id="o-header-drawer-child-{{@index}}" data-trackable="sub-level">
-						{{#each children}}
-						<li class="o-header__drawer-menu-item">
-							<a class="o-header__drawer-menu-link o-header__drawer-menu-link--child o-header__drawer-menu-link--{{#if selected}}selected{{else}}unselected{{/if}}" href="{{href}}" {{#selected}}aria-label="Current page" aria-selected="true"{{/selected}} data-trackable="{{name}}">{{name}}</a>
-						</li>
-						{{/each}}
-					</ul>
-					{{else}}
-					{{#with item}}
-					<a class="o-header__drawer-menu-link o-header__drawer-menu-link--{{../variation}} o-header__drawer-menu-link--{{#if selected}}selected{{else}}unselected{{/if}}" href="{{href}}" {{#selected}}aria-label="Current page" aria-selected="true"{{/selected}} data-trackable="{{name}}">{{name}}</a>
-					{{/with}}
-					{{/if}}
-				</li>
-				{{/if}}
-				{{/each}}
-				{{/each}}
-			</ul>
-		</nav>
-
-		<nav class="o-header__drawer-menu o-header__drawer-menu--user" data-trackable="user-nav">
-			<ul class="o-header__drawer-menu-list">
-				{{#if @root.anon.userIsLoggedIn}}
-				{{#each @root.navigationLists.drawer_bottom.user}}
-				<li class="o-header__drawer-menu-item">
-					<a class="o-header__drawer-menu-link" href="{{href}}" data-trackable="{{name}}">
-						{{name}}
-						{{#if showemail}}
-						<span class="o-header__drawer-menu-link-detail" data-o-header-drawer-user-email></span>
-						{{/if}}
-					</a>
-				</li>
-				{{/each}}
-				{{/if}}
-				{{#if @root.anon.userIsAnonymous}}
-				{{#each @root.navigationLists.drawer_bottom.anon}}
-				<li class="o-header__drawer-menu-item">
-					<a class="o-header__drawer-menu-link" href="{{href}}" data-trackable="{{name}}">
-						{{name}}
-					</a>
-				</li>
-				{{/each}}
-				{{/if}}
-
-				<li class="o-header__drawer-menu-item">
-					{{>n-ui/opt-out/template classname='o-header__drawer-menu-button'}}
-				</li>
-			</ul>
-		</nav>
-	</div>
-</div>
