+++ header/partials/header/anon.html
@@ -2,9 +2,9 @@
 <div class="o-header__row o-header__anon" data-trackable="header-anon">
 	<ul class="o-header__anon-list">
 		{{#each @root.navigationLists.navbar_right.anon}}
-		<li class="o-header__anon-item">
-			<a class="o-header__anon-link" href="{{href}}" data-trackable="{{name}}">{{name}}</a>
-		</li>
+			<li class="o-header__anon-item">
+				<a class="o-header__anon-link" href="{{href}}" data-trackable="{{name}}">{{name}}</a>
+			</li>
 		{{/each}}
 	</ul>
 </div>
