--- header/partials/header/anon_v1.html
@@ -1,11 +1,0 @@
-{{#if @root.anon.userIsAnonymous}}
-	<div class="o-header__row o-header__anon" data-trackable="header-anon">
-		<ul class="o-header__anon-list">
-			{{#each @root.navigation.menus.navbar_right-anon}}
-				<li class="o-header__anon-item">
-					<a class="o-header__anon-link" href="{{url}}" data-trackable="{{label}}">{{label}}</a>
-				</li>
-			{{/each}}
-		</ul>
-	</div>
-{{/if}}
