--- header/partials/header/nav_v2.html
@@ -1,44 +1,0 @@
-{{#ifAll @root.flags.frontPageFastFTJourney @root.navigation.menus.navbar-simple}}
-	<nav id="o-header-nav-mobile" class="o-header__row o-header__nav o-header__nav--mobile" aria-hidden="true" data-trackable="header-nav:mobile">
-		<ul class="o-header__nav-list">
-			{{#each @root.navigation.menus.navbar-simple.items}}
-				<li class="o-header__nav-item">
-					<a class="o-header__nav-link o-header__nav-link--primary" href="{{url}}" {{#selected}}aria-label="Current page" aria-selected="true"{{/selected}} data-trackable="{{name}}">{{label}}</a>
-				</li>
-			{{/each}}
-		</ul>
-	</nav>
-{{/ifAll}}
-
-<nav id="o-header-nav-desktop" class="o-header__row o-header__nav o-header__nav--desktop" role="navigation" aria-label="Primary navigation" data-trackable="header-nav:desktop">
-	<div class="o-header__container">
-		<ul class="o-header__nav-list o-header__nav-list--left" data-trackable="primary-nav">
-			{{#each @root.navigation.menus.navbar.items}}
-			<li class="o-header__nav-item">
-				<a class="o-header__nav-link o-header__nav-link--primary {{#ifEquals @root.currentNav.topLevelRelative.href href}}o-header-nav__list-link--highlight{{/ifEquals}}" href="{{url}}" id="o-header-link-{{@index}}" {{#selected}}aria-label="Current page" aria-selected="true"{{/selected}} data-trackable="{{label}}">{{label}}</a>
-				{{> n-ui/header/partials/meganav/template}}
-			</li>
-			{{/each}}
-		</ul>
-
-
-
-		{{#unlessEquals nUi.header.userNav false}}
-			<ul class="o-header__nav-list o-header__nav-list--right" data-trackable="user-nav">
-				{{#if @root.anon.userIsAnonymous}}
-					{{#each @root.navigation.menus.navbar-right.items}}
-						<li class="o-header__nav-item">
-							<a class="o-header__nav-link" href="{{url}}" data-trackable="{{label}}">{{label}}</a>
-						</li>
-					{{/each}}
-				{{else}}
-					{{#each @root.navigation.menus.navbar-right-anon.items}}
-						<li class="o-header__nav-item">
-							<a class="o-header__nav-link" href="{{url}}" data-trackable="{{label}}">{{label}}</a>
-						</li>
-					{{/each}}
-				{{/if}}
-			</ul>
-		{{/unlessEquals}}
-	</div>
-</nav>
