+++ header/partials/meganav/articlelist.html
@@ -1,31 +1,14 @@
-{{#if @root.flags.origamiNavigation}}
-	<div class="o-header__mega-column o-header__mega-column--articles" data-trackable="popular">
-		<h3 class="o-header__mega-heading">{{title}}</h3>
-		<div class="o-header__mega-content">
-			<ul class="o-header__mega-list">
-				{{#each data}}
-				<li class="o-header__mega-item">
-					<a class="o-header__mega-link" href="{{url}}" data-trackable="link">
-						{{label}}
-					</a>
-				</li>
-				{{/each}}
-			</ul>
-		</div>
+<div class="o-header__mega-column o-header__mega-column--articles" data-trackable="popular">
+	<h3 class="o-header__mega-heading">{{title}}</h3>
+	<div class="o-header__mega-content">
+		<ul class="o-header__mega-list">
+			{{#each data}}
+			<li class="o-header__mega-item">
+				<a class="o-header__mega-link" href="{{href}}" data-trackable="link">
+					{{name}}
+				</a>
+			</li>
+			{{/each}}
+		</ul>
 	</div>
-{{else}}
-	<div class="o-header__mega-column o-header__mega-column--articles" data-trackable="popular">
-		<h3 class="o-header__mega-heading">{{title}}</h3>
-		<div class="o-header__mega-content">
-			<ul class="o-header__mega-list">
-				{{#each data}}
-				<li class="o-header__mega-item">
-					<a class="o-header__mega-link" href="{{href}}" data-trackable="link">
-						{{name}}
-					</a>
-				</li>
-				{{/each}}
-			</ul>
-		</div>
-	</div>
-{{/if}}
+</div>
