--- header/partials/subnav/template_v2.html
@@ -1,41 +1,0 @@
-{{#if @root.navigation.showSubNav}}
-
-	<div class="o-header__subnav" role="navigation" aria-label="Sub navigation" data-o-header-subnav data-trackable="header-subnav">
-		<div class="o-header__container">
-			<div class="o-header__subnav-wrap-outside">
-				<div class="o-header__subnav-wrap-inside" data-o-header-subnav-wrapper>
-					<div class="o-header__subnav-content">
-
-						{{#if @root.navigation.breadcrumb}}
-							<ol class="o-header__subnav-list o-header__subnav-list--breadcrumb" aria-label="Breadcrumb" data-trackable="breadcrumb">
-								{{#each @root.navigation.breadcrumb}}
-								<li class="o-header__subnav-item">
-									<a class="o-header__subnav-link {{#if selected}}o-header__subnav-link--highlight{{/if}}" href="{{url}}" aria-selected="{{selected}}" data-trackable="{{label}}">
-										{{label}}
-									</a>
-								</li>
-								{{/each}}
-							</ol>
-						{{/if}}
-
-						{{#if @root.navigation.subsections}}
-							<ul class="o-header__subnav-list o-header__subnav-list--subsections" aria-label="Subsections" data-trackable="subsections">
-								{{#each @root.navigation.subsections}}
-								<li class="o-header__subnav-item">
-									<a class="o-header__subnav-link {{#if selected}}o-header__subnav-link--highlight{{/if}}" href="{{url}}" aria-selected="{{selected}}" data-trackable="{{label}}">
-										{{label}}
-									</a>
-								</li>
-								{{/each}}
-							</ul>
-						{{/if}}
-
-					</div>
-				</div>
-				<button class="o-header__subnav-button o-header__subnav-button--left" title="scroll left" aria-label="scroll left" aria-hidden="true" disabled></button>
-				<button class="o-header__subnav-button o-header__subnav-button--right" title="scroll right" aria-label="scroll right" aria-hidden="true" disabled></button>
-			</div>
-		</div>
-	</div>
-
-{{/if}}
