--- node/test/stubs/navigationHierarcyWorldUk.json
@@ -1,1 +1,0 @@
-{"item":{"label":"UK","url":"/world/uk","submenu":{"label":null,"items":[{"label":"UK Economy","url":"/global-economy/uk","submenu":null},{"label":"UK Politics & Policy","url":"/world/uk/politics","submenu":null},{"label":"UK Companies","url":"/companies/uk","submenu":null}]}},"siblings":[{"label":"World Economy","url":"/global-economy","submenu":null},{"label":"US","url":"/world/us","submenu":{"label":null,"items":[{"label":"US Economy","url":"/global-economy/us","submenu":null},{"label":"US Politics & Policy","url":"/world/us/politics","submenu":null},{"label":"US Companies","url":"/companies/us","submenu":null},{"label":"US Election 2016","url":"/us-election-2016","submenu":null}]}},{"label":"China","url":"/world/asia-pacific/china","submenu":{"label":null,"items":[{"label":"Chinese Economy","url":"/stream/sectionsId/MTEx-U2VjdGlvbnM=","submenu":null},{"label":"China Politics & Policy","url":"/stream/topicsId/ZTBkYTQyYjMtNjIzYi00NDhkLWIxMDYtOTYwNjdlZDgxZDAx-VG9waWNz","submenu":null}]}},{"label":"Africa","url":"/world/africa","submenu":null},{"label":"Asia Pacific","url":"/world/asia-pacific","submenu":null},{"label":"Emerging Markets","url":"/markets/emerging-markets","submenu":null},{"label":"Europe","url":"/world/europe","submenu":null},{"label":"Latin America","url":"/world/americas","submenu":null},{"label":"Middle East and North Africa","url":"/world/mideast","submenu":null}],"children":[{"label":"UK Economy","url":"/global-economy/uk","submenu":null},{"label":"UK Politics & Policy","url":"/world/uk/politics","submenu":null},{"label":"UK Companies","url":"/companies/uk","submenu":null}],"ancestors":[{"label":"World","url":"/world"}]}
